commit 8 : Ajout JavaDoc Customer

+++ src/main/java/com/rentals/Customer.java
@@ -6,6 +6,9 @@ public class Customer {
     private String name;
     private List<Rental> rentals;
 
+    /** Constructor Class Customer
+     * @param customerName
+     */
     public Customer(final String customerName) {
         super();
         this.name = customerName;
@@ -16,16 +19,25 @@ public class Customer {
         rentals.add(rental);
     }
 
+    /** Getter name
+     * @return name of the customer
+     */
     public String getName() {
         return name;
     }
 
+    /** getter list of rentals
+     * @return list of rentals of the customers
+     */
     public List<Rental> getRentals() {
         return rentals;
     }
 
+    /** Method who create the invoice based by the customers and his rentals
+     * @return a string who contains the amount the customer have to pay
+     */
     public String generateInvoice() {
-        double totalAmount = 0;
+        double totalAmount = (double) 0;
         StringBuilder result = new StringBuilder("Invoice for " + getName() + "\n");
 
         for (Rental rental : rentals) {
